requirements.txt
algokit==1.0.1
anyio==3.6.2
beaker-pyteal==1.0.0b2
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
copier==7.1.0a0
docstring-parser==0.14.1
dunamai==1.16.0
executing==1.2.0
funcy==1.18
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ipfshttpclient==0.7.0
Jinja2==3.1.2
jinja2-ansible-filters==1.3.2
jsonify==0.5
MarkupSafe==2.1.2
msgpack==1.0.5
packaging==23.0
pathspec==0.11.1
plumbum==1.8.1
prompt-toolkit==3.0.38
py-algorand-sdk==2.1.2
pyclip==0.7.0
pycparser==2.21
pycryptodomex==3.17
pydantic==1.10.7
pygame==2.3.0
Pygments==2.14.0
PyNaCl==1.5.0
pyteal==0.23.0
PyYAML==6.0
pyyaml-include==1.3
questionary==1.10.0
rfc3986==1.5.0
semantic-version==2.10.0
shellingham==1.5.0.post1
sniffio==1.3.0
tabulate==0.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
